Made List constructor public and use it from generated code - resolves #141
Made this change to improve the performance of runtime compilation of
predicates when using Java 8. The specific issue that was causing the
runtime compilation in Projog to be slow is that the compiled predicates
used an overloaded version of ListFactory.createList. Generated code now
calls the org.projog.core.term.List constructor directly.
https://bugs.openjdk.java.net/browse/JDK-8145742