revert to interpreted if compilation fails - #175
When operating in "compiled" mode Projog attempts to compile
user-defined predicates to Java at runtime. Made a change so that if the
compilation of a user-defined predicate fails (e.g. because the
generated Java source is too large) then Projog reverts to "interpreted"
mode, and logs a warning, rather than throwing an exception.