add get methods to KnowledgeBase #193
There are a few classes that have a one-to-one relationship with a
KnowledgeBase. These were being managed using the
KnowledgeBaseServiceLocator class. Instead now have instances of these
classes created inside the constructor of the KnowledgeBase and accessed
via get methods. Making this change as it makes it more explicit how
these objects are created and accessed.